# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HUB",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user?key=": {
            "get": {
                "tags": [
                    "/User"
                ],
                "summary": "Returns all available users",
                "operationId": "1041359fda7485c39fcc204add479dbc",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Return the all user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array will be sent with all the found users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user_id": {
                                                "description": "User id, primary key",
                                                "type": "string"
                                            },
                                            "name": {
                                                "description": "Name of the user",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Email of the user registered",
                                                "type": "string"
                                            },
                                            "cpf": {
                                                "description": "CPF of the user registered",
                                                "type": "string"
                                            },
                                            "address": {
                                                "description": "Address of the user registered",
                                                "type": "string"
                                            },
                                            "cep": {
                                                "description": "Cep of the user registered",
                                                "type": "string"
                                            },
                                            "date_birth": {
                                                "description": "Date of birth",
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "status": {
                                                "description": "Status: Active 1 | 0 Deactive",
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "description": "Date the user was created",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Date the user was updated",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error fetching the user"
                    }
                }
            }
        },
...
    }
}